requirements.txt
groq
python-dotenv
pyyaml
streamlit
scikit-learn
matplotlib
